+++ HISTORY.md
@@ -2,6 +2,7 @@
 
 # 0.6.0
 * ENH download metadata template in the GUI
+* ENH multicova analysis
 
 # 0.5.4
 * FIX altair version - binning of streamlit version